Fix CustomEvent in strict mode
If the CustomEvent module is itself used in a strict mode environment
(and all modules run in strict mode in ES6), `window` will be undefined.
This is because the value of `this` is `undefined` in a `null`-reffed
function call in strict mode.

This fix references the `window` object directly.

Fix #6